refactor(procedures): funnel the arch-word return idiom through arch_word (angr-12jjk.23)
~85 sites across 20 files in native/angr/src/procedures/ hand-wrote
`RustBV::concrete(v as u128, state.arch().bits())` (usually preceded by a
`let bits = state.arch().bits();` binding) to build a native SimProcedure's
return value at the ABI return-register width. Adds
`procedures::arch_word(state, value: u64)` next to `check_max`, and rewrites
those sites through it; 45 now-dead `let bits` bindings and 5 now-unused
`RustBV` imports go with them. `declare_const_proc!` expands to the helper too,
so `DEFAULT_UID_GID` drops from u128 to u64.

Behavior-identical: every replaced site's payload was already u64-or-narrower
widened with `as u128`, and the helper's `u128::from(value)` reproduces that
bit pattern exactly (including the `-1i64 as u64` errno returns, which
`RustBV::concrete` still truncates to `bits`).

Left alone on purpose: sites whose value is a `u128` local produced by a
generic scan helper (strset/strstr/strtol) — funnelling those would mean
changing the helper's return type, not removing duplication — and every site
whose width is deliberately *not* the word width (char returns, scanf
conversions sized by their length modifier, symbolic BVs).

+++ native/angr/src/procedures/getid.rs
@@ -12,7 +12,7 @@
 
 /// Angr's default uid/gid return value (mirrors `DEFAULT_UID_GID` in
 /// `syscalls/identity.rs` and `procedures/posix/getuid.py`).
-const DEFAULT_UID_GID: u128 = 1000;
+const DEFAULT_UID_GID: u64 = 1000;
 
 crate::declare_const_proc! {
     /// ```c
